ai-server/requirements.txt
flask
flask-cors
tensorflow
pillow
scikit-learn
numpy